+++ include/Ark/VM/FFI.hpp
@@ -34,6 +34,7 @@ namespace Ark::internal::FFI
     FFI_Function(fileExists);  // fileExists?, 1 argument
     FFI_Function(timeSinceEpoch);  // time, 0 argument
     FFI_Function(sleep);  // sleep, 1 argument
+    FFI_Function(system_);  // system, 1 argument
 }
 
 #undef FFI_Function